Added ```Evaluation Process``` printout feature
