+++ rct229/ruletest_engine/ruletest_jsons/scripts/generate_rule_test_documentation.py
@@ -1,6 +1,5 @@
-from rct229.ruletest_engine.ruletest_jsons.scripts.excel_generation_utilities import (
-    create_rule_test_documentation_spreadsheet,
-)
+from rct229.ruletest_engine.ruletest_jsons.scripts.excel_generation_utilities import \
+    create_rule_test_documentation_spreadsheet
 
 # INPUT:
 ruleset_standard = "ashrae9012019"
